Remove redundant preferences e2e test
